# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Calendar laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/me/password": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update current user password",
                "operationId": "7c960d2b15af12bb5a89ea4203c69cb1",
                "responses": {
                    "200": {
                        "description": "User"
                    },
                    "403": {
                        "description": "Account is registered with google, cannot change password"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}